% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{console_var}
\alias{console_var}
\title{Prints variables in the JS console}
\usage{
console_var(...)
}
\description{
Print one or several variables into the Javascript console. Values will be coerced to character and pasted with `collapse=", "` (see examples).
}
\examples{
#in server.R
console_var(letters[1], letters[2:4])
console_var(names(iris)[1:2], dim(iris))

#output in the JS console:
#letters[1]=[a], letters[2:3]=[b, c, d]
#names(iris)[1:2]=[Sepal.Length, Sepal.Width], dim(iris)=[150, 5]
}
\seealso{
shinyjs::showLog console_log
}
