% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstableAssistant.R
\name{crosstableAssistant}
\alias{crosstableAssistant}
\title{Crosstable Assistant}
\usage{
crosstableAssistant(.data = NULL, viewer = "dialog")
}
\description{
Crosstable Assistant
}
\examples{
crosstableAssistant()
crosstableAssistant(mtcars)
}
