% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enableCheckboxGroupButton}
\alias{enableCheckboxGroupButton}
\title{Enable or disable a button option from a group of `checkboxGroupButtons`}
\usage{
enableCheckboxGroupButton(inputId, value, enable = TRUE)
}
\arguments{
\item{inputId}{the same `inputId` as in `checkboxGroupButtons`}

\item{value}{the value corresponding to the `choices` argument in `checkboxGroupButtons`}

\item{enable}{whether to enable (`TRUE`) or disable (`FALSE`) the button}
}
\value{

}
\description{
Enable or disable a button option from a group of `checkboxGroupButtons`
}
