% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_condition}
\alias{test_condition}
\title{test_condition("is_num", undefined="true")}
\usage{
test_condition(x, go = TRUE, undefined = "true", accessor = "do_show")
}
\description{
test_condition("is_num", undefined="true")
}
